fixes to onChange() and processChar()
Lots of work to get things going properly. The big idea in this refactor branch
was to replace separate signals to the GameFragment and KeyFragment with updates
to a single LiveData object, GameFragment and KeyboardFragment now simply "do the
right thing" instead of being told what to do via signals. In other words, I'm now
using LiveData more as it was intended.